move logic to hf load component only
