+++ components/load_from_hf_hub/src/main.py
@@ -3,6 +3,7 @@ import logging
 import typing as t
 
 import dask.dataframe as dd
+import pandas as pd
 from fondant.component import DaskLoadComponent
 from fondant.executor import DaskLoadExecutor
 
@@ -16,6 +17,7 @@ class LoadFromHubComponent(DaskLoadComponent):
              column_name_mapping: dict,
              image_column_names: t.Optional[list],
              n_rows_to_load: t.Optional[int],
+             index_column:t.Optional[str],
                  ) -> None:
         """
         Args:
@@ -25,11 +27,14 @@ class LoadFromHubComponent(DaskLoadComponent):
                 format the image from HF hub format to a byte string
             n_rows_to_load: optional argument that defines the number of rows to load. Useful for
               testing pipeline runs on a small scale.
+            index_column: Column to set index to in the load component, if not specified a default
+                globally unique index will be set.
         """
         self.dataset_name = dataset_name
         self.column_name_mapping = column_name_mapping
         self.image_column_names = image_column_names
         self.n_rows_to_load = n_rows_to_load
+        self.index_column = index_column
 
     def load(self) -> dd.DataFrame:
         # 1) Load data, read as Dask dataframe
@@ -58,6 +63,28 @@ class LoadFromHubComponent(DaskLoadComponent):
                 partitions_length += len(partition)
             dask_df = dask_df.head(self.n_rows_to_load, npartitions=npartitions)
             dask_df = dd.from_pandas(dask_df, npartitions=npartitions)
+
+        # 4) Set the index
+        if self.index_column is None:
+            logger.info(
+                "Index column not specified, setting a globally unique index",
+            )
+
+            def _set_unique_index(dataframe: pd.DataFrame, partition_info=None):
+                """Function that sets a unique index based on the partition and row number."""
+                dataframe["id"] = 1
+                dataframe["id"] = (
+                    str(partition_info["number"])
+                    + "_"
+                    + (dataframe.id.cumsum()).astype(str)
+                )
+                dataframe.index = dataframe.pop("id")
+                return dataframe
+
+            dask_df = dask_df.map_partitions(_set_unique_index, meta=dask_df.head())
+        else:
+            logger.info(f"Setting `{self.index_column}` as index")
+            dask_df = dask_df.set_index(self.index_column, drop=True)
 
         return dask_df
