+++ examples/pipelines/datacomp/components/load_from_hf_hub/fondant_component.yaml
@@ -50,3 +50,7 @@ args:
     description: Optional argument that defines the number of rows to load. Useful for testing pipeline runs on a small scale
     type: int
     default: None
+  index_column:
+    description: Column to set index to in the load component, if not specified a default globally unique index will be set
+    type: str
+    default: None
