+++ examples/pipelines/starcoder/components/load_from_hub/fondant_component.yaml
@@ -38,3 +38,7 @@ args:
     description: Optional argument that defines the number of rows to load. Useful for testing pipeline runs on a small scale
     type: int
     default: None
+  index_column:
+    description: Column to set index to in the load component, if not specified a default globally unique index will be set
+    type: str
+    default: None
