+++ tests/example_pipelines/compiled_pipeline/example_1/docker-compose.yml
@@ -13,8 +13,6 @@ services:
     - a dummy string arg
     - --input_partition_rows
     - disable
-    - --index_column
-    - None
     - --component_spec
     - '{"name": "First component", "description": "This is an example component",
       "image": "example_component:latest", "produces": {"images": {"fields": {"data":
@@ -35,8 +33,6 @@ services:
     - a dummy string arg
     - --input_partition_rows
     - '10'
-    - --index_column
-    - None
     - --component_spec
     - '{"name": "Second component", "description": "This is an example component",
       "image": "example_component:latest", "consumes": {"images": {"fields": {"data":
@@ -63,8 +59,6 @@ services:
     - --some_list
     - '[1, 2, 3]'
     - --input_partition_rows
-    - None
-    - --index_column
     - None
     - --component_spec
     - '{"name": "Third component", "description": "This is an example component",
