+++ tests/example_pipelines/compiled_pipeline/example_2/docker-compose.yml
@@ -13,8 +13,6 @@ services:
     - a dummy string arg
     - --input_partition_rows
     - None
-    - --index_column
-    - None
     - --component_spec
     - '{"name": "First component", "description": "This is an example component",
       "image": "example_component:latest", "produces": {"images": {"fields": {"data":
@@ -33,8 +31,6 @@ services:
     - --padding
     - '0'
     - --input_partition_rows
-    - None
-    - --index_column
     - None
     - --component_spec
     - '{"name": "Image cropping", "description": "Component that removes single-colored
